ci: format-check every tracked .rs file, not just the include!()d ones
The previous gate derived its file list by grepping for `include!("...")`
and checking those files. That closed the larger half of the hole, but it
made the gate only as correct as the enumeration: a file reached by
neither `mod` nor `include!` is invisible to `cargo fmt --all -- --check`
AND to an include!-derived list.

Handing rustfmt every tracked `.rs` file removes the enumeration from the
trust chain entirely. There is no list to drift, and no walk whose
correctness the gate depends on.

Measured on this commit's parent, one rustfmt invocation over all 3358
tracked `.rs` files:

  - all 3358 clean, 0 flagged, 0 parse failures
  - no rustfmt.toml exists in the tree or above it, so per-file config
    discovery and cargo's agree by construction; the edition is still
    read from [workspace.package] rather than restated

The three unreachable files that motivated this gate were deleted
separately, so it starts from a clean tree. Its value is prospective: it
is what makes the next such file visible instead of silently unchecked.

Control matrix, the same defect planted in three different kinds of file:

  | file kind                    | new gate | cargo fmt --all | old gate |
  |------------------------------|----------|-----------------|----------|
  | reached by `mod`             | exit 1   | exit 1          | exit 0   |
  | reached by `include!()`      | exit 1   | exit 0          | exit 1   |
  | reached by nothing           | exit 1   | exit 0          | exit 0   |

The middle row is the hole the previous gate closed; the last row is the
one it could not see, and it is a property of the file's reachability,
not of any particular file existing today. The first row is the positive
control - without it, `exit 0` from `cargo fmt` would not distinguish a
miss from a pass.

A fragment that only parses at its `include!` site will fail this gate
rather than be skipped. That is deliberate: a skip arm would restore the
same class of invisible gap, and the failure message says what to do.

rustfmt is invoked once with all paths, not once per file; paths stay
relative to the repository root to keep argv well under the platform
limit.
